Glimpse: Reduce view model classes number
