+++ app/src/main/java/org/lineageos/glimpse/ext/AndroidViewModel.kt
@@ -1,0 +1,13 @@
+/*
+ * SPDX-FileCopyrightText: 2023 The LineageOS Project
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.lineageos.glimpse.ext
+
+import android.app.Application
+import android.content.Context
+import androidx.lifecycle.AndroidViewModel
+
+val AndroidViewModel.context: Context
+    get() = getApplication<Application>().applicationContext
