+++ app/src/main/java/org/lineageos/glimpse/viewmodels/AlbumsViewModel.kt
@@ -6,17 +6,19 @@
 package org.lineageos.glimpse.viewmodels
 
 import android.app.Application
+import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
+import org.lineageos.glimpse.ext.*
 import org.lineageos.glimpse.repository.MediaRepository
 
-open class AlbumsViewModel(
+class AlbumsViewModel(
     application: Application,
-) : GlimpseViewModel(application) {
+) : AndroidViewModel(application) {
     val albums = MediaRepository.albums(context).flowOn(Dispatchers.IO).map {
         QueryResult.Data(it)
     }.stateIn(
